+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - none yet
 
 ### Fixed
-- none yet
+- Remove trash folder on offline request
 
 ### Added
 - `cancel()` method on service object returned by `cozy.client.intents.createService()``